+++ src/module/feature/damageHandler/dyingHandling.ts
@@ -11,51 +11,63 @@ export function dyingHandlingPreUpdateActorHook(
     update: Record<string, string>,
     currentActorHp: number,
     updateHp: number,
+    autoGainDying: string,
 ) {
+    const automaticMove = String(game.settings.get(MODULENAME, "enableAutomaticMove"));
     const automoveIfZeroHP =
         game.combat &&
-        ((String(game.settings.get(MODULENAME, "enableAutomaticMove")) === "reaching0HPCharactersOnly" &&
-            actor.type === CHARACTER_TYPE) ||
-            (String(game.settings.get(MODULENAME, "enableAutomaticMove")) === "reaching0HP" &&
-                [CHARACTER_TYPE, NPC_TYPE].includes(actor.type)));
-    if (!String(game.settings.get(MODULENAME, "autoGainDyingAtZeroHP")).startsWith("no")) {
-        handleDyingOnZeroHP(actor, fu.deepClone(update), currentActorHp, updateHp).then((hpRaisedAbove0) => {
-            logDebug("Workbench increaseDyingOnZeroHP complete");
-            if (hpRaisedAbove0) {
-                if (!String(game.settings.get(MODULENAME, "autoRemoveDyingAtGreaterThanZeroHP")).startsWith("no")) {
-                    // Ugh.
-                    new Promise((resolve) => setTimeout(resolve, 250)).then(() => {
-                        autoRemoveDyingAtGreaterThanZeroHp(actor, currentActorHp <= 0 && hpRaisedAbove0).then(() => {
-                            logDebug("Workbench autoRemoveDyingAtGreaterThanZeroHP complete");
-                            if (game.settings.get(MODULENAME, "autoRemoveUnconsciousAtGreaterThanZeroHP")) {
-                                autoRemoveUnconsciousAtGreaterThanZeroHP(
-                                    actor,
-                                    currentActorHp <= 0 && hpRaisedAbove0,
-                                ).then();
-                            }
+        ((automaticMove === "reaching0HPCharactersOnly" && actor.type === CHARACTER_TYPE) ||
+            (automaticMove === "reaching0HP" && [CHARACTER_TYPE, NPC_TYPE].includes(actor.type)));
+    const autoRemoveDying = String(game.settings.get(MODULENAME, "autoRemoveDyingAtGreaterThanZeroHP"));
+    const autoRemoveUnconscious = String(game.settings.get(MODULENAME, "autoRemoveUnconsciousAtGreaterThanZeroHP"));
+
+    if (autoGainDying && !autoGainDying.startsWith("no")) {
+        handleDyingOnZeroHP(actor, fu.deepClone(update), currentActorHp, updateHp, autoGainDying).then(
+            (hpRaisedAbove0) => {
+                logDebug("Workbench increaseDyingOnZeroHP complete");
+                if (hpRaisedAbove0) {
+                    if (autoRemoveDying && !autoRemoveDying.startsWith("no")) {
+                        // Ugh.
+                        new Promise((resolve) => setTimeout(resolve, 250)).then(() => {
+                            autoRemoveDyingAtGreaterThanZeroHp(
+                                actor,
+                                currentActorHp <= 0 && hpRaisedAbove0,
+                                autoRemoveDying,
+                            ).then(() => {
+                                logDebug("Workbench autoRemoveDyingAtGreaterThanZeroHP complete");
+                                if (autoRemoveUnconscious) {
+                                    autoRemoveUnconsciousAtGreaterThanZeroHP(
+                                        actor,
+                                        currentActorHp <= 0 && hpRaisedAbove0,
+                                    ).then();
+                                }
+                            });
                         });
-                    });
+                    } else {
+                        if (autoRemoveUnconscious) {
+                            autoRemoveUnconsciousAtGreaterThanZeroHP(
+                                actor,
+                                currentActorHp <= 0 && hpRaisedAbove0,
+                            ).then();
+                        }
+                    }
                 } else {
-                    if (game.settings.get(MODULENAME, "autoRemoveUnconsciousAtGreaterThanZeroHP")) {
-                        autoRemoveUnconsciousAtGreaterThanZeroHP(actor, currentActorHp <= 0 && hpRaisedAbove0).then();
+                    if (automoveIfZeroHP && currentActorHp > 0 && updateHp <= 0) {
+                        moveOnZeroHP(actor);
                     }
                 }
-            } else {
-                if (automoveIfZeroHP && currentActorHp > 0 && updateHp <= 0) {
-                    moveOnZeroHP(actor);
-                }
-            }
-        });
+            },
+        );
     } else {
         if (currentActorHp <= 0 && updateHp > 0) {
-            if (!String(game.settings.get(MODULENAME, "autoRemoveDyingAtGreaterThanZeroHP")).startsWith("no")) {
-                autoRemoveDyingAtGreaterThanZeroHp(actor, currentActorHp <= 0).then(() => {
-                    if (game.settings.get(MODULENAME, "autoRemoveUnconsciousAtGreaterThanZeroHP")) {
+            if (autoRemoveDying && !autoRemoveDying.startsWith("no")) {
+                autoRemoveDyingAtGreaterThanZeroHp(actor, currentActorHp <= 0, autoRemoveDying).then(() => {
+                    if (autoRemoveUnconscious) {
                         autoRemoveUnconsciousAtGreaterThanZeroHP(actor, currentActorHp <= 0).then();
                     }
                 });
             } else {
-                if (game.settings.get(MODULENAME, "autoRemoveUnconsciousAtGreaterThanZeroHP")) {
+                if (autoRemoveUnconscious) {
                     autoRemoveUnconsciousAtGreaterThanZeroHP(actor, currentActorHp <= 0).then();
                 }
             }
@@ -70,30 +82,29 @@ export async function itemHandlingItemHook(item: ItemPF2e) {
         handleDying(0, 0, item.parent, false);
     }
 
-    if (
-        game.settings.get(MODULENAME, "giveWoundedWhenDyingRemoved") ||
-        game.settings.get(MODULENAME, "giveUnconsciousIfDyingRemovedAt0HP")
-    ) {
-        if (game.settings.get(MODULENAME, "giveWoundedWhenDyingRemoved")) {
+    const giveWounded = game.settings.get(MODULENAME, "giveWoundedWhenDyingRemoved");
+    const giveUnconscious = game.settings.get(MODULENAME, "giveUnconsciousIfDyingRemovedAt0HP");
+    if (giveWounded || giveUnconscious) {
+        if (giveWounded) {
             giveWoundedWhenDyingRemoved(item).then(() => {
                 logDebug("Workbench giveWoundedWhenDyingRemoved complete");
-                if (game.settings.get(MODULENAME, "giveUnconsciousIfDyingRemovedAt0HP")) {
+                if (giveUnconscious) {
                     giveUnconsciousIfDyingRemovedAt0HP(item).then(() => {
                         logDebug("Workbench giveUnconsciousIfDyingRemovedAt0HP complete");
                     });
                 }
             });
-        } else if (game.settings.get(MODULENAME, "giveUnconsciousIfDyingRemovedAt0HP")) {
+        } else if (giveUnconscious) {
             await giveUnconsciousIfDyingRemovedAt0HP(item);
         }
     }
 }
 
-export function handleDyingRecoveryRoll(message: ChatMessagePF2e) {
+export function handleDyingRecoveryRoll(message: ChatMessagePF2e, enabled: boolean) {
     const flavor = message.flavor;
     const token = message.token;
     if (
-        game.settings.get(MODULENAME, "handleDyingRecoveryRoll") &&
+        enabled &&
         shouldIHandleThisMessage(
             message,
             ["all", "players"].includes(String(game.settings.get(MODULENAME, "handleDyingRecoveryRollAllow"))),
@@ -203,13 +214,16 @@ export function handleDying(
     }
 }
 
-export async function autoRemoveDyingAtGreaterThanZeroHp(actor: ActorPF2e, hpAboveZero: boolean): Promise<boolean> {
+export async function autoRemoveDyingAtGreaterThanZeroHp(
+    actor: ActorPF2e,
+    hpAboveZero: boolean,
+    autoRemoveDying,
+): Promise<boolean> {
     const dying = actor.getCondition("dying");
     if (shouldIHandleThis(actor) && dying && !dying.isLocked && hpAboveZero) {
         const value = dying?.value || 0;
         if (dying && value > 0 && !dying.isLocked) {
-            const option = String(game.settings.get(MODULENAME, "autoRemoveDyingAtGreaterThanZeroHP"));
-            if (option.endsWith("ForCharacters") ? ["character", "familiar"].includes(actor.type) : true) {
+            if (autoRemoveDying.endsWith("ForCharacters") ? ["character", "familiar"].includes(actor.type) : true) {
                 handleDying(0, 0, actor);
             }
         }
@@ -378,6 +392,7 @@ export async function handleDyingOnZeroHP(
     update: Record<string, string>,
     hp: number,
     updateHp: number,
+    autogainDying: string,
 ): Promise<boolean> {
     if (!shouldIHandleThis(actor) || hp <= 0 || updateHp > 0) {
         return updateHp > 0;
@@ -388,7 +403,6 @@ export async function handleDyingOnZeroHP(
     let dyingCounter = 0;
     let hpNowAboveZero = false;
     const effectsToCreate: any[] = [];
-    const dyingOption = String(game.settings.get(MODULENAME, "autoGainDyingAtZeroHP"));
     const nonlethalOption = String(game.settings.get(MODULENAME, "nonLethalIsNotLethal"));
 
     const __ret = handleOrcFerocity(actor, update, effectsToCreate, name, shouldIncreaseWounded, hpNowAboveZero);
@@ -399,16 +413,16 @@ export async function handleDyingOnZeroHP(
 
     if (
         !hpNowAboveZero &&
-        (dyingOption.endsWith("ForCharacters") ? ["character", "familiar"].includes(actor.type) : true)
+        (autogainDying.endsWith("ForCharacters") ? ["character", "familiar"].includes(actor.type) : true)
     ) {
-        if (dyingOption?.startsWith("addWoundedLevel")) {
+        if (autogainDying?.startsWith("addWoundedLevel")) {
             dyingCounter = (actor.getCondition("wounded")?.value ?? 0) + 1;
         } else {
             dyingCounter = 1;
         }
     }
 
-    if (checkIfLatestDamageMessageIsCriticalHitByEnemy(actor, dyingOption)) {
+    if (checkIfLatestDamageMessageIsCriticalHitByEnemy(actor, autogainDying)) {
         dyingCounter += 1;
     }
 
@@ -532,7 +546,8 @@ export async function giveUnconsciousIfDyingRemovedAt0HP(item: ItemPF2e) {
 }
 
 export function dyingHandlingPreCreateChatMessageHook(message: ChatMessagePF2e) {
-    if (!String(game.settings.get(MODULENAME, "autoGainDyingIfTakingDamageWhenAlreadyDying")).startsWith("no")) {
+    const autoGainDying = String(game.settings.get(MODULENAME, "autoGainDyingIfTakingDamageWhenAlreadyDying"));
+    if (autoGainDying && !autoGainDying.startsWith("no")) {
         const actor = message.actor;
         if (actor && shouldIHandleThis(actor)) {
             if (message.content?.includes("damage-taken")) {
@@ -543,16 +558,13 @@ export function dyingHandlingPreCreateChatMessageHook(message: ChatMessagePF2e) 
                 // @ts-ignore
                 const notTooSoon = !flag?.between(now - 4000, now);
                 if (notTooSoon) {
-                    const dyingOption = String(
-                        game.settings.get(MODULENAME, "autoGainDyingIfTakingDamageWhenAlreadyDying"),
-                    );
                     const originalDyingCounter = actor?.getCondition("dying")?.value ?? 0;
                     let dyingCounter = 0;
-                    if (!dyingOption.startsWith("no") && originalDyingCounter > 0) {
-                        const wasCritical = checkIfLatestDamageMessageIsCriticalHitByEnemy(actor, dyingOption);
+                    if (!autoGainDying.startsWith("no") && originalDyingCounter > 0) {
+                        const wasCritical = checkIfLatestDamageMessageIsCriticalHitByEnemy(actor, autoGainDying);
 
                         if (
-                            dyingOption.endsWith("ForCharacters")
+                            autoGainDying.endsWith("ForCharacters")
                                 ? ["character", "familiar"].includes(actor.type)
                                 : true
                         ) {